+++ Production/govuk_ios/Model/RecentActivity/RecentActivtityHompageCell.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct RecentActivityHomepageCell {
+    let title: String
+    let lastVisitedString: String
+}
